Arreglos
Navbar arreglado. Sigue sin funcionar la sincronización del counter
